balance type changed from double to BigDecimal, method exchangeCurrency amended

+++ src/main/java/com/adrian/bank/account/AccountController.java
@@ -3,6 +3,8 @@ package com.adrian.bank.account;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RestController
 public class AccountController {
 
@@ -11,7 +13,7 @@ public class AccountController {
 
 
     @PostMapping("/account")
-    public Account makeAccount(String name, double balance, Currency currency) {
+    public Account makeAccount(String name, BigDecimal balance, Currency currency) {
         return accountService.createAccount (name, balance, currency);
     }
 
@@ -32,17 +34,17 @@ public class AccountController {
     }
 
     @PostMapping("/account/deposit")
-    public BalanceResponse deposit(long id, double depositSum) throws Exception {
+    public BalanceResponse deposit(long id, BigDecimal depositSum) throws Exception {
         return accountService.deposit (id, depositSum);
     }
 
     @PostMapping("/account/withdrawal")
-    public BalanceResponse withdrawal(long id, double withdrawalsum) throws Exception {
+    public BalanceResponse withdrawal(long id, BigDecimal withdrawalsum) throws Exception {
         return accountService.withdrawal (id, withdrawalsum);
     }
 
     @PostMapping("account/transfer")
-    public BalanceResponse transfer(long idFromAcc, long idToAcc, double transferSum) throws Exception {
+    public BalanceResponse transfer(long idFromAcc, long idToAcc, BigDecimal transferSum) throws Exception {
         return accountService.transferFunds (idFromAcc, idToAcc, transferSum);
     }
